requirements.txt
Flask
Pillow
opencv-python
numpy
ngrok
psutil
